refactor(game): extract GameListService out of game_service facade
